requirements.txt
Django==2.2.10
djangorestframework~=3.12.2
psycopg2-binary>=2.8